requirements.txt
# sudo apt install libsox-dev
torch
torchaudio
numpy
scipy
einops
pyyaml
huggingface_hub
encodec==0.1.1
tensorboardX
tensorboard
matplotlib
transformers
soundfile
jsonargparse[signatures]
librosa
torchcrepe
pesq
